separate styles from jsx files
