+++ src/components/Auth/AuthForm/AuthFormStyle.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  title: {
+    height: 50,
+    padding: 5,
+    fontSize: 30,
+    fontWeight: "bold"
+  }
+});
+
+
+export default styles;
