+++ src/components/Auth/AuthText/AuthTextStyle.js
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  account: {
+    display: "flex",
+    flexDirection: "row"
+  },
+  accountText: {
+    color: "rgba(107, 104, 255, 1)",
+    fontWeight: "bold"
+  }
+});
+
+
+export default styles;
