+++ src/components/UI/Auth/AuthButtonStyle.js
@@ -1,0 +1,22 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  button: {
+    height: 50,
+    width: 350,
+    backgroundColor: "rgba(107, 104, 255, 1)",
+    borderColor: "#000000",
+    borderWidth: 1,
+    marginBottom: 10,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  buttonText: {
+    color: "white",
+    fontSize: 20,
+    fontWeight: "bold"
+  },
+})
+
+
+export default styles;
