+++ src/components/UI/Auth/AuthInput.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
-import { TextInput, StyleSheet } from "react-native";
+import { TextInput } from "react-native";
+import styles from "./AuthInputStyle";
 
 function AuthInput(props) {
   const [isSecure, setIsSecure] = useState(false);
@@ -20,19 +21,5 @@ function AuthInput(props) {
     />
   );
 }
-
-const styles = StyleSheet.create({
-  input: {
-    height: 50,
-    width: 350,
-    marginBottom: 10,
-    borderBottomColor: "rgba(107, 104, 255, 1)",
-    borderBottomWidth: 1,
-    borderRadius: 2,
-    padding: 15,
-    shadowRadius: 1.41,
-    elevation: 0.1
-  }
-});
 
 export default AuthInput;
