+++ src/components/UI/Auth/AuthInputStyle.js
@@ -1,0 +1,17 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  input: {
+    height: 50,
+    width: 350,
+    marginBottom: 10,
+    borderBottomColor: "rgba(107, 104, 255, 1)",
+    borderBottomWidth: 1,
+    borderRadius: 2,
+    padding: 15,
+    shadowRadius: 1.41,
+    elevation: 0.1
+  }
+});
+
+export default styles;
