+++ src/pages/AuthPageStyle.js
@@ -1,0 +1,20 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: "#fff",
+    alignItems: "center",
+    margin: 0,
+    padding: 0
+  },
+
+  background: {
+    marginTop: 0,
+    width: 1080,
+    height: 280,
+    resizeMode: "contain",
+    position: "relative"
+  }
+});
+
+export default styles;
